+++ Changelog.md
@@ -5,6 +5,7 @@ The changes for each individual (beta) release can be seen [here](https://github
 # 2.2.2.0
 
 - Added `ROTATOR-MOVED` and `ROTATOR-MOVED-MECHANICAL` events to the websocket
+- Added `application/get-tab` to get the current application tab
 
 ## 2.2.1.0